backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-dotenv
Pillow
numpy

# Roboflow Inference (for hosted inference)
# Docs: https://inference.roboflow.com
# Note: Requires ROBOFLOW_API_KEY environment variable
inference
inference-sdk
supervision
